cleanup and separation of code
