+++ services/ui-src/src/hooks/authHooks/useUser.tsx
@@ -1,0 +1,13 @@
+import { useContext } from "react";
+import { UserContext } from "./userContext";
+
+export const useUser = () => {
+  const context = useContext(UserContext);
+
+  if (context === undefined) {
+    throw new Error(
+      "`useUser` hook must be used within a `UserProvider` component"
+    );
+  }
+  return context;
+};
