+++ services/ui-src/src/hooks/authHooks/userContext.tsx
@@ -1,0 +1,18 @@
+import { createContext } from "react";
+import { CognitoUser } from "@aws-amplify/auth";
+
+export interface UserContextInterface {
+  user?: CognitoUser;
+  showLocalLogins?: boolean;
+  logout: () => Promise<void>;
+  loginWithIDM: () => Promise<void>;
+}
+
+export const UserContext = createContext<UserContextInterface>({
+  logout: async () => {
+    console.log("User Context failed to initialize logout functionality");
+  },
+  loginWithIDM: async () => {
+    console.log("User Context failed to initialize IDM login functionality.");
+  },
+});
